Updated Digis and analysis module

+++ RecoDataProducts/inc/STMWaveformDigi.hh
@@ -20,17 +20,17 @@ namespace mu2e {
 
     STMWaveformDigi() : _DetID(-1), _trigTimeOffset(0), _adcs(std::vector<int16_t>()), _peakpos(0){};
 
-    STMWaveformDigi(int DetID, uint32_t trigTimeOffset, std::vector<int16_t> adcs, int peakpos) : _DetID(DetID), _trigTimeOffset(trigTimeOffset), _adcs(adcs), _peakpos(peakpos) {};
-    STMWaveformDigi(int DetID, uint32_t trigTimeOffset, std::vector<int16_t> adcs) : _DetID(DetID), _trigTimeOffset(trigTimeOffset), _adcs(adcs), _peakpos(0) {};
+    STMWaveformDigi(int16_t DetID, uint32_t trigTimeOffset, std::vector<int16_t> adcs, int peakpos) : _DetID(DetID), _trigTimeOffset(trigTimeOffset), _adcs(adcs), _peakpos(peakpos) {};
+    STMWaveformDigi(int16_t DetID, uint32_t trigTimeOffset, std::vector<int16_t> adcs) : _DetID(DetID), _trigTimeOffset(trigTimeOffset), _adcs(adcs), _peakpos(0) {};
     STMWaveformDigi(uint32_t trigTimeOffset, std::vector<int16_t> adcs) : _trigTimeOffset(trigTimeOffset), _adcs(adcs) {};
 
-    int                     DetID()   const {return _DetID;}
+    int16_t                     DetID()   const {return _DetID;}
     uint32_t trigTimeOffset() const { return _trigTimeOffset; }
     const std::vector<int16_t>& adcs() const { return _adcs; }
     int                     peakpos()  const {return _peakpos;}
 
   private:
-    int _DetID;
+    int16_t _DetID;
     uint32_t _trigTimeOffset; // time offset from EWT? to first ADC value [ct]
     std::vector<int16_t> _adcs; // vector of ADC values for the waveform
     int _peakpos;
